refactor : throttlingManager infra 부분으로 이동(#72)
